+++ src/test/java/com/minionz/backend/shop/controller/ShopControllerTest.java
@@ -162,9 +162,9 @@ class ShopControllerTest extends ApiDocument {
         list.add(new ShopTableCountResponseDto(2, 2));
         list.add(new ShopTableCountResponseDto(3, 5));
         list.add(new ShopTableCountResponseDto(4, 7));
-        double maxUserA = 47;
-        double liveUser = 5;
-        ShopMaxUserResponseDto maxUser = new ShopMaxUserResponseDto(maxUserA, liveUser, (liveUser / maxUserA) * 100);
+        int maxUserA = 47;
+        int liveUser = 5;
+        ShopMaxUserResponseDto maxUser = new ShopMaxUserResponseDto(maxUserA, liveUser, (liveUser / (double)maxUserA) * 100);
         List<ShopDetailsResponseDto> shopDetailsResponseDto = new ArrayList<>();
         shopDetailsResponseDto.add(new ShopDetailsResponseDto(name, address, telNumber, list, id, maxUser));
         willReturn(shopDetailsResponseDto).given(shopService).viewDetails(any(Long.class));
@@ -176,7 +176,6 @@ class ShopControllerTest extends ApiDocument {
     @Test
     void 매장_상세보기_조회_실패() throws Exception {
         Long id = 1L;
-        ;
         Message message = new Message("매장 상세보기 조회 실패");
         willThrow(new NotFoundException("매장 상세보기 조회 실패")).given(shopService).viewDetails(any(Long.class));
         ResultActions resultActions = 유저_매장_상세보기_조회_요청(id);
